Move processRestart queue to separate class, reduce in coordinator size #4051
