+++ AlphaWallet/InCoordinator.swift
@@ -491,10 +491,6 @@ class InCoordinator: NSObject, Coordinator {
         tabBarController.viewControllers = viewControllers
     }
 
-    @objc private func dismissTransactions() {
-        navigationController.dismiss(animated: true)
-    }
-
     func showTab(_ selectTab: UITabBarController.Tabs) {
         guard let viewControllers = tabBarController.viewControllers else {
             return
@@ -508,18 +504,9 @@ class InCoordinator: NSObject, Coordinator {
         }
     }
 
-    private func removeAllCoordinators() {
-        coordinators.removeAll()
-    }
-
     private func checkDevice() {
-        let deviceChecker = CheckDeviceCoordinator(
-                navigationController: navigationController,
-                jailbreakChecker: DeviceChecker()
-        )
-
+        let deviceChecker = CheckDeviceCoordinator(navigationController: navigationController, jailbreakChecker: DeviceChecker())
         deviceChecker.start()
-
         addCoordinator(deviceChecker)
     }
 
@@ -651,73 +638,12 @@ class InCoordinator: NSObject, Coordinator {
     }
 
     private func processRestartQueueAndRestartUI() {
-        processRestartQueueBeforeRestart(config: config, restartQueue: restartQueue)
+        RestartQueueHandler(config: config).processRestartQueueBeforeRestart(restartQueue: restartQueue)
         restartUI(withReason: .serverChange, account: wallet)
     }
 
     private func restartUI(withReason reason: RestartReason, account: Wallet) {
         delegate?.didRestart(in: self, reason: reason, wallet: account)
-    }
-
-    private func processRestartQueueBeforeRestart(config: Config, restartQueue: RestartTaskQueue) {
-        for each in restartQueue.queue {
-            switch each {
-            case .addServer(let server):
-                restartQueue.remove(each)
-                RPCServer.customRpcs.append(server)
-            case .editServer(let original, let edited):
-                restartQueue.remove(each)
-                replaceServer(original: original, edited: edited)
-            case .removeServer(let server):
-                restartQueue.remove(each)
-                removeServer(server)
-            case .enableServer(let server):
-                restartQueue.remove(each)
-                var c = config
-                // NOTE: we need to make sure that we don't enableServer test net server when main net is selected.
-                // update enabledServers with added server
-                var servers = c.enabledServers.filter({ $0.isTestnet == server.isTestnet })
-                servers.append(server)
-                c.enabledServers = servers
-            case .switchDappServer(server: let server):
-                restartQueue.remove(each)
-                Config.setChainId(server.chainID)
-            case .loadUrlInDappBrowser:
-                break
-            case .reloadServers(let servers):
-                restartQueue.remove(each)
-                var c = config
-                c.enabledServers = servers
-            }
-        }
-    }
-
-    private func replaceServer(original: CustomRPC, edited: CustomRPC) {
-        RPCServer.customRpcs = RPCServer.customRpcs.map { (item: CustomRPC) -> CustomRPC in
-            if item.chainID == original.chainID {
-                return edited
-            }
-            return item
-        }
-    }
-
-    private func removeServer(_ server: CustomRPC) {
-        //Must disable server first because we (might) not have done that if the user had disabled and then remove the server in the UI at the same time. And if we fallback to mainnet when an enabled server's chain ID is not found, this can lead to mainnet appearing twice in the Wallet tab
-        let servers = config.enabledServers.filter { $0.chainID != server.chainID }
-        var config = self.config
-        config.enabledServers = servers
-        guard let i = RPCServer.customRpcs.firstIndex(of: server) else { return }
-        RPCServer.customRpcs.remove(at: i)
-        switchBrowserServer(awayFrom: server, config: config)
-    }
-
-    private func switchBrowserServer(awayFrom server: CustomRPC, config: Config) {
-        guard Config.getChainId() == server.chainID else { return }
-        //To be safe, we find a network that is either mainnet/testnet depending on the chain that was removed
-        let isTestnet = server.isTestnet
-        if let targetServer = config.enabledServers.first(where: { $0.isTestnet == isTestnet }) {
-            Config.setChainId(targetServer.chainID)
-        }
     }
 
     private func processRestartQueueAfterRestart(config: Config, coordinator: InCoordinator, restartQueue: RestartTaskQueue) {
